Split native hit target audit collaborators
